# swagger-output.json
{
    "openapi":  "3.0.4",
    "info":  {
                 "title":  "AI-PMS.API",
                 "version":  "1.0"
             },
    "paths":  {
...
                  "/api/BottleneckDetection/project/{projectId}":  {
                                                                       "post":  {
                                                                                    "tags":  [
                                                                                                 "BottleneckDetection"
                                                                                             ],
                                                                                    "parameters":  [
                                                                                                       {
                                                                                                           "name":  "projectId",
                                                                                                           "in":  "path",
                                                                                                           "required":  true,
                                                                                                           "schema":  {
                                                                                                                          "type":  "string",
                                                                                                                          "format":  "uuid"
                                                                                                                      }
                                                                                                       }
                                                                                                   ],
                                                                                    "responses":  {
                                                                                                      "200":  {
                                                                                                                  "description":  "OK",
                                                                                                                  "content":  {
                                                                                                                                  "text/plain":  {
                                                                                                                                                     "schema":  {
                                                                                                                                                                    "$ref":  "#/components/schemas/BottleneckResponse"
                                                                                                                                                                }
                                                                                                                                                 },
                                                                                                                                  "application/json":  {
                                                                                                                                                           "schema":  {
                                                                                                                                                                          "$ref":  "#/components/schemas/BottleneckResponse"
                                                                                                                                                                      }
                                                                                                                                                       },
                                                                                                                                  "text/json":  {
                                                                                                                                                    "schema":  {
                                                                                                                                                                   "$ref":  "#/components/schemas/BottleneckResponse"
                                                                                                                                                               }
                                                                                                                                                }
                                                                                                                              }
                                                                                                              }
                                                                                                  }
                                                                                }
                                                                   },
                  "/api/DeadlinePrediction":  {
                                                  "post":  {
                                                               "tags":  [
                                                                            "DeadlinePrediction"
                                                                        ],
                                                               "requestBody":  {
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/DeadlinePredictionRequest"
                                                                                                                                       }
                                                                                                                        },
                                                                                                   "text/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/DeadlinePredictionRequest"
                                                                                                                                }
                                                                                                                 },
                                                                                                   "application/*+json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/DeadlinePredictionRequest"
                                                                                                                                         }
                                                                                                                          }
                                                                                               }
                                                                               },
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  "OK"
                                                                                         }
                                                                             }
                                                           }
                                              },
                  "/api/DeadlinePrediction/project/{projectId}":  {
                                                                      "post":  {
                                                                                   "tags":  [
                                                                                                "DeadlinePrediction"
                                                                                            ],
                                                                                   "parameters":  [
                                                                                                      {
                                                                                                          "name":  "projectId",
                                                                                                          "in":  "path",
                                                                                                          "required":  true,
                                                                                                          "schema":  {
                                                                                                                         "type":  "string",
                                                                                                                         "format":  "uuid"
                                                                                                                     }
                                                                                                      }
                                                                                                  ],
                                                                                   "responses":  {
                                                                                                     "200":  {
                                                                                                                 "description":  "OK"
                                                                                                             }
                                                                                                 }
                                                                               }
                                                                  },
                  "/api/ProjectProgress":  {
                                               "post":  {
                                                            "tags":  [
                                                                         "ProjectProgress"
                                                                     ],
                                                            "requestBody":  {
                                                                                "content":  {
                                                                                                "application/json":  {
                                                                                                                         "schema":  {
                                                                                                                                        "$ref":  "#/components/schemas/ProjectProgressRequest"
                                                                                                                                    }
                                                                                                                     },
                                                                                                "text/json":  {
                                                                                                                  "schema":  {
                                                                                                                                 "$ref":  "#/components/schemas/ProjectProgressRequest"
                                                                                                                             }
                                                                                                              },
                                                                                                "application/*+json":  {
                                                                                                                           "schema":  {
                                                                                                                                          "$ref":  "#/components/schemas/ProjectProgressRequest"
                                                                                                                                      }
                                                                                                                       }
                                                                                            }
                                                                            },
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "OK"
                                                                                      }
                                                                          }
                                                        }
                                           },
                  "/api/ProjectProgress/project/{projectId}":  {
                                                                   "post":  {
                                                                                "tags":  [
                                                                                             "ProjectProgress"
                                                                                         ],
                                                                                "parameters":  [
                                                                                                   {
                                                                                                       "name":  "projectId",
                                                                                                       "in":  "path",
                                                                                                       "required":  true,
                                                                                                       "schema":  {
                                                                                                                      "type":  "string",
                                                                                                                      "format":  "uuid"
                                                                                                                  }
                                                                                                   }
                                                                                               ],
                                                                                "responses":  {
                                                                                                  "200":  {
                                                                                                              "description":  "OK"
                                                                                                          }
                                                                                              }
                                                                            }
                                                               },
...
                  "/api/TeamPerformance/project/{projectId}":  {
                                                                   "post":  {
                                                                                "tags":  [
                                                                                             "TeamPerformance"
                                                                                         ],
                                                                                "parameters":  [
                                                                                                   {
                                                                                                       "name":  "projectId",
                                                                                                       "in":  "path",
                                                                                                       "required":  true,
                                                                                                       "schema":  {
                                                                                                                      "type":  "string",
                                                                                                                      "format":  "uuid"
                                                                                                                  }
                                                                                                   }
                                                                                               ],
                                                                                "responses":  {
                                                                                                  "200":  {
                                                                                                              "description":  "OK"
                                                                                                          }
                                                                                              }
                                                                            }
                                                               }
              },
    "components":  {
                       "schemas":  {
...
                                       "BottleneckItem":  {
                                                              "type":  "object",
                                                              "properties":  {
                                                                                 "taskTitle":  {
                                                                                                   "type":  "string",
                                                                                                   "nullable":  true
                                                                                               },
                                                                                 "reason":  {
                                                                                                "type":  "string",
                                                                                                "nullable":  true
                                                                                            },
                                                                                 "severity":  {
                                                                                                  "type":  "string",
                                                                                                  "nullable":  true
                                                                                              },
                                                                                 "suggestedActions":  {
                                                                                                          "type":  "array",
                                                                                                          "items":  {
                                                                                                                        "type":  "string"
                                                                                                                    },
                                                                                                          "nullable":  true
                                                                                                      }
                                                                             },
                                                              "additionalProperties":  false
                                                          },
...
                                       "BottleneckResponse":  {
                                                                  "type":  "object",
                                                                  "properties":  {
                                                                                     "bottlenecks":  {
                                                                                                         "type":  "array",
                                                                                                         "items":  {
                                                                                                                       "$ref":  "#/components/schemas/BottleneckItem"
                                                                                                                   },
                                                                                                         "nullable":  true
                                                                                                     },
                                                                                     "recommendations":  {
                                                                                                             "type":  "array",
                                                                                                             "items":  {
                                                                                                                           "type":  "string"
                                                                                                                       },
                                                                                                             "nullable":  true
                                                                                                         }
                                                                                 },
                                                                  "additionalProperties":  false
                                                              },
...
                                       "DeadlinePredictionRequest":  {
                                                                         "type":  "object",
                                                                         "properties":  {
                                                                                            "projectName":  {
                                                                                                                "type":  "string",
                                                                                                                "nullable":  true
                                                                                                            },
                                                                                            "projectDeadline":  {
                                                                                                                    "type":  "string",
                                                                                                                    "format":  "date-time",
                                                                                                                    "nullable":  true
                                                                                                                },
                                                                                            "totalTasks":  {
                                                                                                               "type":  "integer",
                                                                                                               "format":  "int32"
                                                                                                           },
                                                                                            "completedTasks":  {
                                                                                                                   "type":  "integer",
                                                                                                                   "format":  "int32"
                                                                                                               },
                                                                                            "remainingTasks":  {
                                                                                                                   "type":  "integer",
                                                                                                                   "format":  "int32"
                                                                                                               },
                                                                                            "inProgressTasks":  {
                                                                                                                    "type":  "integer",
                                                                                                                    "format":  "int32"
                                                                                                                },
                                                                                            "blockedTasks":  {
                                                                                                                 "type":  "integer",
                                                                                                                 "format":  "int32"
                                                                                                             },
                                                                                            "overdueTasks":  {
                                                                                                                 "type":  "integer",
                                                                                                                 "format":  "int32"
                                                                                                             },
                                                                                            "totalEstimatedHours":  {
                                                                                                                        "type":  "integer",
                                                                                                                        "format":  "int32"
                                                                                                                    },
                                                                                            "totalActualHours":  {
                                                                                                                     "type":  "integer",
                                                                                                                     "format":  "int32"
                                                                                                                 },
                                                                                            "completionRate":  {
                                                                                                                   "type":  "number",
                                                                                                                   "format":  "double"
                                                                                                               }
                                                                                        },
                                                                         "additionalProperties":  false
                                                                     },
...
                                   }
                   },
...
}